[SSCP][libkernel] Remove stddef.h and size_t from more places
